+++ tests/check_stringTables.c
@@ -32,9 +32,9 @@ START_TEST(test_createValueTable) {
 
     fail_unless(err == EXIP_OK, "createValueTable returns error code %d", err);
     fail_unless(valueTable.count == 0,
-            "createValueTable populates the valueTable with count: %d", valueTable.count);
+            "createValueTable populates the valueTable with count: %d", (int)valueTable.count);
     fail_unless(valueTable.dynArray.arrayEntries == DEFAULT_VALUE_ENTRIES_NUMBER,
-            "createValueTable creates dynamic array with %d rows", valueTable.dynArray.arrayEntries);
+            "createValueTable creates dynamic array with %d rows", (int)valueTable.dynArray.arrayEntries);
     fail_if(valueTable.value == NULL);
 
     destroyDynArray(&valueTable.dynArray);
@@ -62,7 +62,7 @@ START_TEST(test_addUriEntry) {
     fail_unless(stringEqual(uriTable.uri[0].uriStr, test_uri) == 1,
             "addUriEntry changed the uriStr");
     fail_unless(entryId == 0,
-            "addUriEntry returned wrong entryId: %d", entryId);
+            "addUriEntry returned wrong entryId: %d", (int)entryId);
 
     fail_if(uriTable.uri[0].lnTable.ln == NULL);
 
@@ -78,7 +78,7 @@ START_TEST(test_addUriEntry) {
     fail_unless(stringEqual(uriTable.uri[DEFAULT_URI_ENTRIES_NUMBER].uriStr, test_uri) == 1,
             "addUriEntry changed the uriStr");
     fail_unless(entryId == DEFAULT_URI_ENTRIES_NUMBER,
-            "addUriEntry returned wrong entryId: %d", entryId);
+            "addUriEntry returned wrong entryId: %d", (int)entryId);
 
     fail_if(uriTable.uri[DEFAULT_URI_ENTRIES_NUMBER].lnTable.ln == NULL);
     
@@ -114,7 +114,7 @@ START_TEST(test_addLnEntry) {
     fail_unless(stringEqual(lnTable.ln[0].lnStr, test_ln) == 1,
             "addLnEntry changed the lnStr");
     fail_unless(entryId == 0,
-            "addLnEntry returned wrong entryId: %d", entryId);
+            "addLnEntry returned wrong entryId: %d", (int)entryId);
 
 #if VALUE_CROSSTABLE_USE
     fail_if(lnTable.ln[0].vxTable != NULL);
@@ -132,7 +132,7 @@ START_TEST(test_addLnEntry) {
     fail_unless(stringEqual(lnTable.ln[DEFAULT_LN_ENTRIES_NUMBER].lnStr, test_ln) == 1,
             "addLnEntry changed the lnStr");
     fail_unless(entryId == DEFAULT_LN_ENTRIES_NUMBER,
-            "addLnEntry returned wrong entryId: %d", entryId);
+            "addLnEntry returned wrong entryId: %d", (int)entryId);
 #if VALUE_CROSSTABLE_USE
     fail_if(lnTable.ln[DEFAULT_LN_ENTRIES_NUMBER].vxTable != NULL);
 #endif
@@ -185,7 +185,7 @@ START_TEST(test_addValueEntry) {
 
     // Freeing the string tables
     int i;
-    for (i = 0; i < testStrm.schema->uriTable.count; i++) {
+    for (i = 0; i < (int)testStrm.schema->uriTable.count; i++) {
 
         destroyDynArray(&testStrm.schema->uriTable.uri[i].pfxTable.dynArray);
         destroyDynArray(&testStrm.schema->uriTable.uri[i].lnTable.dynArray);
